requirements.txt
accelerate==1.6.0
diffusers==0.33.1
transformers==4.46.2
huggingface_hub>=0.20.2
sentencepiece==0.2.0
pillow==11.1.0
av==12.1.0
numpy==1.26.4
scipy==1.12.0
requests==2.31.0
torchsde==0.2.6
einops
opencv-contrib-python
safetensors
